Use uint64 for VU IDs

+++ core/local/local.go
@@ -160,7 +160,7 @@ func (e *ExecutionScheduler) initVU(
 	// Get the VU ID here, so that the VUs are (mostly) ordered by their
 	// number in the channel buffer
 	vuID := e.state.GetUniqueVUIdentifier()
-	vu, err := e.runner.NewVU(int64(vuID), samplesOut)
+	vu, err := e.runner.NewVU(vuID, samplesOut)
 	if err != nil {
 		return nil, errext.WithHint(err, fmt.Sprintf("error while initializing VU #%d", vuID))
 	}
